backend/requirements.txt
# Core Flask and extensions
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
pytest==8.3.3
pytest-flask==1.3.0

# Database (SQLite is built-in, no extra driver needed)
SQLAlchemy==2.0.36

# OpenAI API (for generating meal/mood suggestions)
openai==1.51.0

# Utilities
python-dotenv==1.0.1
requests==2.32.3
